+++ template/{{cookiecutter.year}}/src/day02.rs
@@ -1,8 +1,8 @@
-pub fn day02p1(_input: &str) -> usize {
+pub fn day02p1(_input: &str) -> u32 {
     0
 }
 
-pub fn day02p2(_input: &str) -> usize {
+pub fn day02p2(_input: &str) -> u32 {
     0
 }
