+++ template/{{cookiecutter.year}}/src/day13.rs
@@ -1,8 +1,8 @@
-pub fn day13p1(_input: &str) -> usize {
+pub fn day13p1(_input: &str) -> u32 {
     0
 }
 
-pub fn day13p2(_input: &str) -> usize {
+pub fn day13p2(_input: &str) -> u32 {
     0
 }
